### 2018-05-16 #732 @leeper
@lebebr01 Would you be interested in reviewing this for JOSS?
