### 2018-05-16 #739 @whedon
[REVIEW]: rsimsum: Summarise results from Monte Carlo simulation studies
**Submitting author:** @ellessenne (<a href="http://orcid.org/0000-0002-8319-7624">Alessandro Gasparini</a>)
**Repository:** <a href="https://github.com/ellessenne/rsimsum/" target ="_blank">https://github.com/ellessenne/rsimsum/</a>
**Version:** 0.3.1
