### 2018-05-16 #732 @leeper
@lebebr01 Great. Thank you! The review issue and instructions are open at: https://github.com/openjournals/joss-reviews/issues/739 Please follow the checklist and let me know (on that thread) if you have any questions.
