### 2018-06-05 42c6a2e
Added hex sticker
