+++ NEWS.md
@@ -1,3 +1,5 @@
+# rsimsum 0.3.2-9000
+
 # rsimsum 0.3.2
 
 `rsimsum` is a small maintenance release:
