### 2018-06-06 ellessenne/rsimsum#2 @lebebr01
Equations not rendering in Vignettes
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/739).

The equations are currently not rendering in any of the vignettes. This occurred when I used `install.packages("rsimsum")` and installing directly from GitHub with `devtools::install_github("ellessenne/rsimsum", build_vignettes = TRUE)`.

This makes the introduction vignette particularly difficult and also makes it difficult to understand the computations the package does without users digging into the code.
