### 2018-06-06 ellessenne/rsimsum#3 @lebebr01
Clarifications within Introduction vignette
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/739).

Below are some items I feel would benefit from some additional clarification in the vignette to ensure users are comfortable with using the `simsum` function and understanding the output.

- [ ] The data from the first example are described well, however there was very little discussion of how to use the `simsum` function except for showing the function used with the MIsim data. For example, how do we know the true value is 0.5 or -0.5, why is `ref = "CC"` and not specified in the second example, etc?
- [ ] I found the discussion of the output from `simsum` to be relatively brief which leaves a lot of burden on the user to understand the output coming from the function with no discussion. I believe some discussion of the output would improve the user experience and confidence in understanding how to interpret the specific statistics.
- [ ] I wonder if a table would be useful for what the values in the "stat" column represent after running the `get_data` function.
- [ ] From the vignette and `simsum` documentation, it is not completely clear to me the difference between the `methodvar` and `by` arguments. I believe the main distinction seems to be within vs between simulation conditions (this was gathered from the JOSS paper), however from the documentation itself, I was unsure.
