### 2018-06-06 ellessenne/rsimsum#4 @lebebr01
Plotting functions
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/739).

A few questions/clarifications I had regarding the plotting output and functions.

- [ ] The `pattern` and `zip` plot functions do not need to specify the `by` argument, but the other plotting functions do. Is it possible to reuse the `by` argument from the `simsum` output object? This would be more consistent across the plotting functions and I also think would be easier for users to not have to reuse the same arguments across functions.
- [ ] Is the sample size order in the figures customizable? In some of the figures from the plotting vignette they go from smallest to largest, but others go from largest to smallest. It'd be great if this could be customizable, otherwise being consistent would be second best.
- [ ] Specific to the plotting vignette. I understand what "multiple estimands" means for the `multsimsum` function, however to ensure the vignette is more approachable to a variety of audiences, I would consider adding synonyms for the word estimands used in other fields.
