### 2018-06-06 ellessenne/rsimsum#5 @lebebr01
Statement of Need
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/739).

I found the statement of need well articulated in the JOSS paper itself, however I wonder if this would be useful to also articulate in the README of the software repository. Users are probably more likely to see the package README first compared to the JOSS paper, therefore being more clear about the uses and goals of the package in the README is beneficial in my mind.
