### 2018-06-06 #739 @lebebr01
@leeper I have completed my review and have posted a few issues for @ellessenne to consider. Many of the issues are related to clarification of arguments, output, or documentation of the functions. I'm happy to look over the package again after @ellessenne has a chance to consider the issues.
