### 2018-06-06 3134ab3
Improved print method of summary.simsum and summary.multisimsum objects
