### 2018-06-06 06c0b50
Fixed rounding error in sprintf introduced by 3134ab38b5a7278979497e81aca93ab7125d842d
