+++ R/print.summary.simsum.R
@@ -77,8 +77,8 @@ print.summary.simsum <- function(x, digits = 4, sstat = "all", ...) {
   names(x$summ)[names(x$summ) == "est"] <- "Estimate"
   if (x$mcse) {
     names(x$summ)[names(x$summ) == "mcse"] <- "MCSE"
-    names(x$summ)[names(x$summ) == "lower"] <- paste("Lower", sprintf("%.0f%%", 100 * (1 - x$ci_level) / 2))
-    names(x$summ)[names(x$summ) == "upper"] <- paste("Upper", sprintf("%.0f%%", 100 - 100 * (1 - x$ci_level) / 2))
+    names(x$summ)[names(x$summ) == "lower"] <- paste("Lower", sprintf("%.1f%%", 100 * (1 - x$ci_level) / 2))
+    names(x$summ)[names(x$summ) == "upper"] <- paste("Upper", sprintf("%.1f%%", 100 * (1 - (1 - x$ci_level) / 2)))
   }
 
   ### Print pretty summary
