### 2018-06-06 b773f67
Improved package entry in R manual
