+++ NAMESPACE
@@ -42,3 +42,6 @@ export(nsim)
 export(pattern)
 export(simsum)
 export(zip)
+import(checkmate)
+import(ggplot2)
+import(stats)
