### 2018-06-06 abfc67e
Added 'description' argument to each get_data method
