+++ R/get_data.summary.multisimsum.R
@@ -17,6 +17,7 @@
 #' * `cover`, coverage of a nominal `level`\% confidence interval.
 #' * `bccover`, bias corrected coverage of a nominal `level`\% confidence interval.
 #' * `power`, power of a (1 - `level`)\% level test.#'
+#' @param description Append a column to the returned `data.frame` with a description of each summary statistic? Defaults to `FALSE`.
 #' @param ... Ignored.
 #' @return A `data.frame` containing summary statistics from a simulation study.
 #' @seealso [summary.multisimsum()], [get_data()]
@@ -29,13 +30,17 @@
 #'    by = "fv_dist")
 #' sms <- summary(ms)
 #' get_data(sms)
+#' get_data(sms, description = TRUE)
 
-get_data.summary.multisimsum <- function(x, sstat = "all", ...) {
+get_data.summary.multisimsum <- function(x, sstat = "all", description = FALSE, ...) {
   ### Check arguments
   arg_checks <- checkmate::makeAssertCollection()
 
   # `sstat` must be one of the possible choices
   checkmate::assert_subset(sstat, choices = c("all", "nsim", "thetamean", "thetamedian", "se2mean", "se2median", "bias", "empse", "mse", "relprec", "modelse", "relerror", "cover", "bccover", "power"), add = arg_checks)
+
+  # `description` must be a boolean value
+  checkmate::assert_logical(description, add = arg_checks)
 
   ### Report if there are any errors
   if (!arg_checks$isEmpty()) checkmate::reportAssertions(arg_checks)
@@ -43,6 +48,11 @@ get_data.summary.multisimsum <- function(x, sstat = "all", ...) {
   ### Select only summary statistics on interest
   if (!("all" %in% sstat)) {
     x$summ <- x$summ[x$summ$stat %in% sstat, ]
+  }
+
+  ### Add description column if requested
+  if (description) {
+  	x$summ <- merge_description(x$summ, ref = x$ref, level = x$level)
   }
 
   ### Return data
