### 2018-06-06 21c0008
Improved clarity of documentation and introductory vignette
