### 2018-06-06 ellessenne/rsimsum#3 @ellessenne
Hello,
I pushed some updates to improve documentation and the introductory vignette, hopefully the points you highlighted are more clear now!
I also added a `description` argument to each `get_data` method: if set to `TRUE`, an additional column with a description of each summary statistic is attached to the `data.frame` return by the function.
What do you think about these changes?
