### 2018-06-07 ellessenne/rsimsum#2 @ellessenne
Hi,
I can't seem to reproduce this problem - I tried on several machines (macOS High Sierra, Windows 7, Windows 10), and it generally works fine for me.
Does the problem occur within RStudio? I sometimes find the RStudio viewer a bit fiddly myself, as sometimes css and js are not loaded properly. If so, could you try the following on the plain R console (not within RStudio):

```r
vignette(topic = "introduction", package = "rsimsum")
```

Does MathJax load the equations this way?
