### 2018-06-11 ellessenne/rsimsum#2 @lebebr01
That is the issue, the equations do not render in RStudio. That is good to know, I wasn't aware of that limitation.

This issue may be worth mentioning in the README as RStudio is becoming common.
