### 2018-06-11 ellessenne/rsimsum#2 @ellessenne
I know, unfortunately there is nothing I can do about it - there is an open issue with RStudio [here](https://github.com/rstudio/rstudio/issues/2253), hopefully this should be resolved soon!
In the meanwhile, I can sure add a little "disclaimer" within the README file!
