### 2018-06-11 ellessenne/rsimsum#3 @lebebr01
I like the changes. I find the additional clarification in the introduction vignette and documentation to be very helpful and I can imagine would be useful for users new to simulation studies.

Secondly, I really like the description argument to the `get_data` function. Very useful and unique solution, better than referencing a table in the documentation.
