### 2018-06-11 ellessenne/rsimsum#4 @ellessenne
Hello,
here are my comments point by point:
1. `pattern` and `zip` plot estimated values and standard errors directly, while the remaining plotting functions plot summary statistics; therefore, I thought it was more interesting to plot every combination of data-generating mechanisms and methods for the former, while requesting to manually specify `by` factors manually for the latter. I am not sure inferring `by` factors would be wise, as (for instance) Monte Carlo simulation studies with several data-generating mechanisms would return plots that are too messy to be useful. Regardless, it is possible to call functions such as `lolly` and `forest` without specifying any `by` factor, and a relevant warning will be thrown suggesting to manually applying faceting as desired to obtain a sensible plot;
2. All values of data-generating mechanisms are by default stored as strings - I will add a section to the plotting vignette on how to manually rearrange and customise the appearance of plots;
3. Thanks, I will try to clarify - any suggestion regarding synonyms?
