### 2018-06-11 bb5b382
Updated README.Rmd with warning for RStudio users [skip ci]
