+++ README.md
@@ -1,16 +1,28 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-rsimsum <img src="man/figures/hex.png" width = "150" align="right" />
-=====================================================================
 
-[![Travis-CI Build Status](https://travis-ci.org/ellessenne/rsimsum.svg?branch=master)](https://travis-ci.org/ellessenne/rsimsum) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ellessenne/rsimsum?branch=master&svg=true)](https://ci.appveyor.com/project/ellessenne/rsimsum) [![Coverage Status](https://img.shields.io/codecov/c/github/ellessenne/rsimsum/master.svg)](https://codecov.io/github/ellessenne/rsimsum?branch=master) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/rsimsum)](https://cran.r-project.org/package=rsimsum) [![CRAN\_Logs\_Badge](http://cranlogs.r-pkg.org/badges/rsimsum)](https://cran.r-project.org/package=rsimsum) [![CRAN\_Logs\_Badge\_Total](http://cranlogs.r-pkg.org/badges/grand-total/rsimsum)](https://cran.r-project.org/package=rsimsum) [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
+# rsimsum <img src="man/figures/hex.png" width = "150" align="right" />
 
-`rsimsum` is an R package for summarising results from simulation studies and computing Monte Carlo standard errors of commonly used summary statistics.
+[![Travis-CI Build
+Status](https://travis-ci.org/ellessenne/rsimsum.svg?branch=master)](https://travis-ci.org/ellessenne/rsimsum)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/ellessenne/rsimsum?branch=master&svg=true)](https://ci.appveyor.com/project/ellessenne/rsimsum)
+[![Coverage
+Status](https://img.shields.io/codecov/c/github/ellessenne/rsimsum/master.svg)](https://codecov.io/github/ellessenne/rsimsum?branch=master)
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/rsimsum)](https://cran.r-project.org/package=rsimsum)
+[![CRAN\_Logs\_Badge](http://cranlogs.r-pkg.org/badges/rsimsum)](https://cran.r-project.org/package=rsimsum)
+[![CRAN\_Logs\_Badge\_Total](http://cranlogs.r-pkg.org/badges/grand-total/rsimsum)](https://cran.r-project.org/package=rsimsum)
+[![PRs
+Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg)](http://makeapullrequest.com)
 
-This package is modelled on the `simsum` user-written command in `Stata` (White I.R., 2010).
+`rsimsum` is an R package for summarising results from simulation
+studies and computing Monte Carlo standard errors of commonly used
+summary statistics.
 
-Installation
-------------
+This package is modelled on the `simsum` user-written command in `Stata`
+(White I.R., 2010).
+
+## Installation
 
 You can install `rsimsum` from CRAN:
 
@@ -18,17 +30,19 @@ You can install `rsimsum` from CRAN:
 install.packages("rsimsum")
 ```
 
-Alternatively, it is possible to install the development version from GitHub via:
+Alternatively, it is possible to install the development version from
+GitHub via:
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("ellessenne/rsimsum")
 ```
 
-Example
--------
+## Example
 
[…]
