+++ README.Rmd
@@ -24,7 +24,7 @@ options(width = 100)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/rsimsum)](https://cran.r-project.org/package=rsimsum)
 [![CRAN_Logs_Badge](http://cranlogs.r-pkg.org/badges/rsimsum)](https://cran.r-project.org/package=rsimsum)
 [![CRAN_Logs_Badge_Total](http://cranlogs.r-pkg.org/badges/grand-total/rsimsum)](https://cran.r-project.org/package=rsimsum)
-[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
+[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg)](http://makeapullrequest.com)
 
 `rsimsum` is an R package for summarising results from simulation studies and computing Monte Carlo standard errors of commonly used summary statistics. 
 
@@ -107,6 +107,14 @@ vignette(topic = "plotting", package = "rsimsum")
 
 * White, I.R. 2010. _simsum: Analyses of simulation studies including Monte Carlo error_. The Stata Journal 10(3): 369-385 <[http://www.stata-journal.com/article.html?article=st0200](http://www.stata-journal.com/article.html?article=st0200)>
 * Morris, T.P, White, I.R. and Crowther, M.J. 2017. _Using simulation studies to evaluate statistical methods_. <[arXiv:1712.03198](https://arxiv.org/abs/1712.03198)>
+
+# Warning for RStudio users
+
+If you use RStudio and equations are not displayed properly within the RStudio viewer window, please access the vignette from [the CRAN website](https://cran.rstudio.com/web/packages/rsimsum/) or directly from the R console with the command:
+
+```r
+vignette(topic = "introduction", package = "rsimsum")
+```
 
 # Copyright
 
