### 2018-06-11 5c9ee00
Updated plotting vignette
