### 2018-06-11 ab377ae
Re-styled a couple of files
