+++ R/get_data.multisimsum.R
@@ -50,7 +50,7 @@ get_data.multisimsum <- function(x, sstat = "all", description = FALSE, ...) {
 
   ### Add description column if requested
   if (description) {
-  	x$summ <- merge_description(x$summ, ref = x$ref, level = x$level)
+    x$summ <- merge_description(x$summ, ref = x$ref, level = x$level)
   }
 
   ### Return data
