### 2018-06-12 ellessenne/rsimsum#5 @ellessenne
Hello, I just updated the README file on GitHub by adding bits and bobs from the JOSS paper!
