### 2018-06-12 #739 @ellessenne
Hello @lebebr01 and @leeper,
I just finished making some changes following @lebebr01 comments (for which I am very grateful - thank you very much!). Do you think there are other things I could improve on?
