### 2018-06-12 66b11ff
Updated vignettes warning for RStudio users
