### 2018-06-12 0adcde1
Improved testing coverage
