### 2018-06-12 9fcd052
Further improved testing coverage
