### 2018-06-13 31874bd
Replaced sapply for vapply in tests
