### 2018-06-14 cdae526
Removed date from DESCRIPTION
