### 2018-06-19 ellessenne/rsimsum#4 @lebebr01
These comments make sense to me. Thanks for articulating.

Regarding 3, perhaps parameter or theta would be a synonyms?
