### 2018-06-19 #739 @lebebr01
@leeper I took a look again and LGTM.

@ellessenne thank you for the comments and further clarification of the documentation.
