### 2018-06-20 #739 @leeper
Thank you, @lebebr01!

@ellessenne Please create a release of the package and issue a DOI for it using, for example, figshare or zenodo, and let me know the DOI once you have it.
