### 2018-06-20 #739 @ellessenne
Thank you @lebebr01 for the review!

@leeper I just archived the current version of `rsimsum` using Zenodo and the corresponding DOI is: [10.5281/zenodo.1293527](http://doi.org/10.5281/zenodo.1293527)!
