### 2018-06-20 ellessenne/rsimsum#4 @ellessenne
Thanks, I am adding "parameters" to the vignette and an additional line explaining what it means in that specific example!
