### 2018-06-20 451ca94
Updated vignette on plotting
